Separate issue report from cleanup service
Put the report code into its own service and handle
the dependency chain from within the systemd unit
files. report runs after all services which creates a
status report and cleanup runs after report
