+++ azure_li_services/units/cleanup.py
@@ -15,12 +15,8 @@
 # You should have received a copy of the GNU General Public License
 # along with azure-li-services.  If not, see <http://www.gnu.org/licenses/>
 #
-import os
-from collections import OrderedDict
-
 # project
 from azure_li_services.command import Command
-from azure_li_services.status_report import StatusReport
 
 
 def main():
@@ -28,35 +24,7 @@ def main():
     Azure Li/Vli cleanup
 
     Uninstall azure-li-services package and its dependencies
-    Report overall service status in an update to /etc/issue
-    if not all services were successful
     """
-    unit_to_service_map = {
-        'config_lookup':
-            'azure-li-config-lookup',
-        'user':
-            'azure-li-user',
-        'install':
-            'azure-li-install',
-        'network':
-            'azure-li-network',
-        'call':
-            'azure-li-call'
-    }
-    unit_to_service_map_ordered = OrderedDict(
-        sorted(unit_to_service_map.items())
-    )
-    failed_services = []
-    for unit, service in unit_to_service_map_ordered.items():
-        report = StatusReport(unit, init_state=False)
-        report.load()
-        success = report.get_state()
-        if not success:
-            failed_services.append(service)
-
-    if failed_services:
-        report_services_failed(failed_services)
-
     Command.run(
         [
             'zypper', '--non-interactive',
@@ -64,17 +32,3 @@ def main():
             'azure-li-services'
         ]
     )
-
-
-def report_services_failed(failed_services):
-    with open('/etc/issue', 'w') as issue:
-        issue.write(os.linesep)
-        issue.write('!!! DEPLOYMENT ERROR !!!')
-        issue.write(os.linesep)
-        issue.write(
-            'For details see: "systemctl status {0}"'.format(
-                ' '.join(failed_services)
-            )
-        )
-        issue.write(os.linesep)
-        issue.write(os.linesep)
