+++ systemd/azure-li-report.service
@@ -1,0 +1,13 @@
+[Unit]
+Description=Report status of Azure Li/VLi services
+ConditionPathExists=/.azure-li-report.trigger
+After=azure-li-config-lookup.service azure-li-call.service azure-li-install.service azure-li-network.service azure-li-user.service
+Before=azure-li-cleanup.service
+
+[Service]
+Type=oneshot
+ExecStart=/usr/bin/azure-li-report
+ExecStartPost=/bin/bash -c 'rm -f /.azure-li-report.trigger'
+
+[Install]
+WantedBy=multi-user.target
